finish back

+++ back/src/RaceRoute.Core/Migrations/20231212012030_InitialCreate.cs
@@ -17,7 +17,7 @@ namespace RaceRoute.Core.Migrations
                     Id = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(max)", nullable: true),
-                    Height = table.Column<int>(type: "int", nullable: false)
+                    Height = table.Column<double>(type: "float", nullable: false)
                 },
                 constraints: table =>
                 {
