+++ back/src/RaceRoute.Core/Migrations/RaceRouteDbContextModelSnapshot.cs
@@ -29,8 +29,8 @@ namespace RaceRoute.Core.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<int>("Height")
-                        .HasColumnType("int");
+                    b.Property<double>("Height")
+                        .HasColumnType("float");
 
                     b.Property<string>("Name")
                         .HasColumnType("nvarchar(max)");
